Player sprite can now move left and right

+++ BurgerTime/Component.cpp
@@ -6,7 +6,7 @@ Component::Component(Engine* engine, Entity* entity) {
 }
 
 void Component::init() { }
-void Component::update(float dt) { }
+void Component::update(double dt) { }
 void Component::receive(Message message) { }
 
 Component::~Component() { }
